backend/fastapi_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.1
aiomysql==0.2.0
PyMySQL==1.1.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
celery==5.3.4
boto3==1.29.0
python-decouple==3.8
pytest==7.4.3
pytest-asyncio==0.21.1
